Add null_hasher, reorganize crate
LruCache itself hashes keys so it can put them in the eviction queue in
a compact way, rather than doing something like storing the key itself.
However, LruCache is backed by a plain old HashMap, which takes this
hashed value, a u64, and hashes it again.

To avoid the double-hashing, this new NullHasher accepts u64's (and only
u64) and simply returns that exact u64 as the thing it hashed - doing
nothing. There's some bit tomfoolery as the u64 is broken into bytes
then reassembled, but it's no more groady than it was before.