Loosen access types
This brings the Cache contract more closely into alignment with HashMap,
which it is modeled off of. HashMap, for get, get_mut, and remove, uses
a type Q rather than K to allow a little more fungibility in those
functions use. This can have some surprising effects, for example see
how the test case had to be fixed. But most people are probably using
Strings for their cache keys, so this effectively makes the differences
between &str and String a little less violent.