Refactor to a LinkedHashMap
This is slightly better because it's one less data structure to break.
The previous implementation was desynchronizing between the eviction
queue and the data storage map itself. This implementation lays the
eviction queue itself onto the data storage map, so removal goes through
the LinkedHashMap which makes it far easier to keep things consistent.
You can see the consistency check in the testing code, in fact.

My hope is this will make it feature-stable enough to re-introduce to
mysteriousbot.